Killed unused sqlinsert.go and tests. Removed remaining reference to deleted batcher in test util code.
